refactoring Pkcs7, Pkcs12, X509Store, X509StoreCtx and X509VerifyParam

+++ lib/utils/ObjectHolder.h
@@ -67,7 +67,7 @@ class ObjectHolder
 
     void release()
     {
-        if (m_isAcquired)
+        if (m_isAcquired && m_raw)
             Type::destroy(m_raw);
 
         m_raw = nullptr;
